configstore: move logic from db to action handler
Some logic lived in the db package, move it to the action handler.
